+++ test/gui/test_auto_tracker_window.py
@@ -52,7 +52,8 @@ async def test_on_timer_update(current_status: GameConnectionStatus, correct_gam
     window._update_timer.start.assert_called_once_with()
 
 
-@pytest.mark.parametrize("name", ["Metroid Prime 1 - Game Art (Standard)", "Metroid Prime 2 - Game Art (Standard)"])
+@pytest.mark.parametrize("name", ["Metroid Prime - Game Art (Standard)",
+                                  "Metroid Prime 2: Echoes - Game Art (Standard)"])
 def test_create_tracker(window: AutoTrackerWindow, name):
     window.create_tracker()
     assert type(window.item_tracker) is QtWidgets.QWidget