+++ test/gui/test_game_session_window.py
@@ -82,6 +82,7 @@ async def test_on_game_session_meta_update(preset_manager, skip_qtbot):
     await window.on_game_session_meta_update(second_session)
     window.update_multiworld_client_status.assert_awaited()
     network_client.session_self_update.assert_awaited_once_with(
+        game_connection.expected_game,
         game_connection.get_current_inventory.return_value,
         game_connection.current_status,
         game_connection.backend_choice,